+++ chain/network/src/stats/metrics.rs
@@ -1,6 +1,6 @@
 use crate::types::PeerMessage;
 use near_metrics::{
-    inc_counter_by_opt, inc_counter_opt, try_create_histogram, try_create_int_counter,
+    do_create_int_counter_vec, try_create_histogram, try_create_int_counter,
     try_create_int_counter_vec, try_create_int_gauge, Histogram, IntCounter, IntCounterVec,
     IntGauge,
 };
@@ -111,9 +111,12 @@ pub static PARTIAL_ENCODED_CHUNK_REQUEST_DELAY: Lazy<Histogram> = Lazy::new(|| {
         .unwrap()
 });
 
-#[derive(Clone)]
+#[derive(Clone, Debug, actix::MessageResponse)]
 pub struct NetworkMetrics {
-    pub peer_messages: HashMap<String, Option<IntCounter>>,
+    // received messages
+    peer_messages: HashMap<String, IntCounter>,
+    // sent messages (broadcast style)
+    pub broadcast_messages: IntCounterVec,
 }
 
 impl NetworkMetrics {
@@ -129,13 +132,20 @@ impl NetworkMetrics {
                         NetworkMetrics::peer_message_bytes_rx,
                         NetworkMetrics::peer_message_dropped,
                     ]
-                    .map(|method| {
+                    .into_iter()
+                    .filter_map(|method| {
                         let counter_name = method(name);
-                        let counter = try_create_int_counter(&counter_name, &counter_name).ok();
-                        (counter_name, counter)
+                        try_create_int_counter(&counter_name, &counter_name)
+                            .ok()
+                            .map(|counter| (counter_name, counter))
                     })
                 })
                 .collect(),
+            broadcast_messages: do_create_int_counter_vec(
+                "near_broadcast_msg",
+                "Broadcasted messages",
+                &["type"],
+            ),
         }
     }
 
@@ -153,13 +163,16 @@ impl NetworkMetrics {
 
     pub fn inc(&self, message_name: &str) {
         if let Some(counter) = self.peer_messages.get(message_name) {
-            inc_counter_opt(counter.as_ref());
+            counter.inc();
         }
     }
 
     pub fn inc_by(&self, message_name: &str, value: u64) {
         if let Some(counter) = self.peer_messages.get(message_name) {
-            inc_counter_by_opt(counter.as_ref(), value);
+            counter.inc_by(value);
         }
+    }
+    pub fn inc_broadcast(&self, message_name: &str) {
+        self.broadcast_messages.with_label_values(&[message_name]).inc();
     }
 }