+++ core/primitives/src/runtime/config_store.rs
@@ -22,7 +22,8 @@ static CONFIGS: &[(ProtocolVersion, &[u8])] = &[
     (50, include_config!("50.json")),
     // max_gas_burnt increased to 300 TGas
     (52, include_config!("52.json")),
-    // Increased deployment costs, increased wasmer2 stack_limit
+    // Increased deployment costs, increased wasmer2 stack_limit, added limiting of contract locals,
+    // set read_cached_trie_node cost, decrease storage key limit
     (53, include_config!("53.json")),
 ];
 
@@ -91,8 +92,10 @@ impl RuntimeConfigStore {
 mod tests {
     use super::*;
     use crate::serialize::to_base;
-    use crate::version::ProtocolFeature::LowerDataReceiptAndEcrecoverBaseCost;
     use crate::version::ProtocolFeature::LowerStorageCost;
+    use crate::version::ProtocolFeature::{
+        LowerDataReceiptAndEcrecoverBaseCost, LowerStorageKeyLimit,
+    };
     use near_primitives_core::hash::hash;
 
     const GENESIS_PROTOCOL_VERSION: ProtocolVersion = 29;
@@ -124,7 +127,7 @@ mod tests {
             "B8fL27CPSdug9AQXgS77b3n9aWFUFf7WUjJcuRMvjE6C",
             "5SneeY6PB8NktERjNq8Gge68cdEgjELiAbhNsmDv1cqY",
             "EFq13cxe78LdB7bbvYF5s1PYxsDGU1teXYXW2qaJkXwW",
-            "611aXAik5tGfHzKNjLsPEB8NQNdJ2q6JGDQCNcFQN3ra",
+            "ENnHEU1otcAWjxv4HhNkKb7FsZFZr5FGNhFjRL2W2a68",
         ];
         let actual_hashes = CONFIGS
             .iter()
@@ -225,6 +228,17 @@ If you add new config version, add a missing hash to the end of `expected_hashes
         assert!(
             base_cfg.wasm_config.ext_costs.ecrecover_base
                 > new_cfg.wasm_config.ext_costs.ecrecover_base
+        );
+    }
+
+    #[test]
+    fn test_lower_max_length_storage_key() {
+        let store = RuntimeConfigStore::new(None);
+        let base_cfg = store.get_config(LowerStorageKeyLimit.protocol_version() - 1);
+        let new_cfg = store.get_config(LowerStorageKeyLimit.protocol_version());
+        assert!(
+            base_cfg.wasm_config.limit_config.max_length_storage_key
+                > new_cfg.wasm_config.limit_config.max_length_storage_key
         );
     }
 }