+++ core/store/src/trie/shard_tries.rs
@@ -15,7 +15,7 @@ use near_primitives::types::{
 use crate::trie::trie_storage::{TrieCache, TrieCachingStorage};
 use crate::trie::{TrieRefcountChange, POISONED_LOCK_ERR};
 use crate::{DBCol, DBOp, DBTransaction};
-use crate::{StorageError, Store, StoreUpdate, Trie, TrieChanges, TrieUpdate};
+use crate::{Store, StoreUpdate, Trie, TrieChanges, TrieUpdate};
 
 struct ShardTriesInner {
     store: Store,
@@ -112,7 +112,7 @@ impl ShardTries {
         tries: ShardTries,
         shard_uid: ShardUId,
         store_update: &mut StoreUpdate,
-    ) -> Result<(), StorageError> {
+    ) {
         store_update.tries = Some(tries);
         for TrieRefcountChange { trie_node_or_value_hash, trie_node_or_value, rc } in
             deletions.iter()
@@ -128,7 +128,6 @@ impl ShardTries {
                 -(*rc as i64),
             );
         }
-        Ok(())
     }
 
     fn apply_insertions_inner(
@@ -136,7 +135,7 @@ impl ShardTries {
         tries: ShardTries,
         shard_uid: ShardUId,
         store_update: &mut StoreUpdate,
-    ) -> Result<(), StorageError> {
+    ) {
         store_update.tries = Some(tries);
         for TrieRefcountChange { trie_node_or_value_hash, trie_node_or_value, rc } in
             insertions.iter()
@@ -152,7 +151,6 @@ impl ShardTries {
                 *rc as i64,
             );
         }
-        Ok(())
     }
 
     fn apply_all_inner(
@@ -160,23 +158,23 @@ impl ShardTries {
         tries: ShardTries,
         shard_uid: ShardUId,
         apply_deletions: bool,
-    ) -> Result<(StoreUpdate, StateRoot), StorageError> {
+    ) -> (StoreUpdate, StateRoot) {
         let mut store_update = StoreUpdate::new_with_tries(tries.clone());
         ShardTries::apply_insertions_inner(
             &trie_changes.insertions,
             tries.clone(),
             shard_uid,
             &mut store_update,
-        )?;
+        );
         if apply_deletions {
             ShardTries::apply_deletions_inner(
                 &trie_changes.deletions,
                 tries,
                 shard_uid,
                 &mut store_update,
-            )?;
+            );
         }
-        Ok((store_update, trie_changes.new_root))
+        (store_update, trie_changes.new_root)
     }
 
     pub fn apply_insertions(
@@ -184,7 +182,7 @@ impl ShardTries {
         trie_changes: &TrieChanges,
         shard_uid: ShardUId,
         store_update: &mut StoreUpdate,
-    ) -> Result<(), StorageError> {
+    ) {
         ShardTries::apply_insertions_inner(
             &trie_changes.insertions,
             self.clone(),
@@ -198,7 +196,7 @@ impl ShardTries {
         trie_changes: &TrieChanges,
         shard_uid: ShardUId,
         store_update: &mut StoreUpdate,
-    ) -> Result<(), StorageError> {
+    ) {
         ShardTries::apply_deletions_inner(
             &trie_changes.deletions,
             self.clone(),
@@ -212,7 +210,7 @@ impl ShardTries {
         trie_changes: &TrieChanges,
         shard_uid: ShardUId,
         store_update: &mut StoreUpdate,
-    ) -> Result<(), StorageError> {
+    ) {
         ShardTries::apply_deletions_inner(
             &trie_changes.insertions,
             self.clone(),
@@ -225,7 +223,7 @@ impl ShardTries {
         &self,
         trie_changes: &TrieChanges,
         shard_uid: ShardUId,
-    ) -> Result<(StoreUpdate, StateRoot), StorageError> {
+    ) -> (StoreUpdate, StateRoot) {
         ShardTries::apply_all_inner(trie_changes, self.clone(), shard_uid, true)
     }
 
@@ -280,7 +278,7 @@ impl WrappedTrieChanges {
         &self.state_changes
     }
 
-    pub fn insertions_into(&self, store_update: &mut StoreUpdate) -> Result<(), StorageError> {
+    pub fn insertions_into(&self, store_update: &mut StoreUpdate) {
         self.tries.apply_insertions(&self.trie_changes, self.shard_uid, store_update)
     }