+++ core/store/src/trie/trie_storage.rs
@@ -232,7 +232,8 @@ impl TrieCachingStorage {
 
 impl TrieStorage for TrieCachingStorage {
     fn retrieve_raw_bytes(&self, hash: &CryptoHash) -> Result<Arc<[u8]>, StorageError> {
-        // Try to get value from chunk cache containing free of charge nodes.
+        // Try to get value from chunk cache containing nodes with cheaper access. We can do it for any `TrieCacheMode`,
+        // because we charge for reading nodes only when `CachingChunk` mode is enabled anyway.
         if let Some(val) = self.chunk_cache.borrow_mut().get(hash) {
             self.inc_mem_read_nodes();
             return Ok(val.clone());