+++ tools/restored-receipts-verifier/src/main.rs
@@ -1,6 +1,6 @@
 use clap::{Arg, Command};
 use near_chain::{ChainStore, ChainStoreAccess, RuntimeAdapter};
-use near_chain_configs::GenesisValidationMode;
+use near_chain_configs::{GenesisValidationMode, DEFAULT_GC_NUM_EPOCHS_TO_KEEP};
 use near_primitives::hash::CryptoHash;
 use near_primitives::receipt::Receipt;
 use near_store::create_store;
@@ -60,6 +60,7 @@ fn main() -> Result<()> {
         None,
         near_config.client_config.max_gas_burnt_view,
         None,
+        DEFAULT_GC_NUM_EPOCHS_TO_KEEP,
     );
 
     let mut receipts_missing = Vec::<Receipt>::new();